Added model for handling unknown events
This model will allow easier extension by end users for complex/simple custom events as it provides all the required data rather than just the name. Also moved the project type to be .net standard 2.0 which should be fine for most consumers but is more indicative of it being a library not an application itself.