Added support for store redemption, also fixed bug with provider check
This contains the main change for store redemption support, so when someone redeems something from the store it will allow you to hook into that event.

There was also an issue where for some reason the complex ones kept blowing up when testing the redemptions as the provider field was missing, so this has been fixed so if it doesnt exist just short circuit. Also I have moved the jwt to be on the connection not the creation of the instance as it is only needed at point of connecting.